Apply global filter for affected-by groups
