+++ addons/promote/common/src/main/resources/default-promote.conf
@@ -11,7 +11,3 @@
 # promote rules and rule-sets.
 #
 #basedir=promote
-
-# Ignore matched affected groups when doing clean-up for the promotion
-#
-#affected.groups.exclude=^build-\d+
