+++ core/src/main/resources/default-internal-features.conf
@@ -2,3 +2,7 @@
 # By default, we disable ArtifactStore validation when repos are stored / updated. This is a feature flag for that
 # behavior, and we'll change this value when we change the default value in the InternalFeatureConfig class.
 #store.validation.enabled=false
+
+# Ignore affected groups when doing clean-up for the promotion
+#
+#affected.groups.exclude=^build-\d+
