+++ db/common/src/main/java/org/commonjava/indy/db/common/AbstractStoreDataManager.java
@@ -61,6 +61,8 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static java.util.Collections.emptySet;
+import static org.apache.commons.lang.StringUtils.isBlank;
 import static org.commonjava.indy.db.common.StoreUpdateAction.DELETE;
 import static org.commonjava.indy.db.common.StoreUpdateAction.STORE;
 import static org.commonjava.indy.model.core.StoreType.group;
@@ -558,7 +560,31 @@ public abstract class AbstractStoreDataManager
             }
         }
 
-        return groups;
+        return filterAffectedGroups( groups );
+    }
+
+    /**
+     * Filter unnecessary affected groups in clean-up process. Most likely to exclude all the temp groups.
+     */
+    protected Set<Group> filterAffectedGroups( Set<Group> affectedGroups )
+    {
+        if ( affectedGroups == null )
+        {
+            return emptySet();
+        }
+        if ( internalFeatureConfig == null )
+        {
+            return affectedGroups;
+        }
+        String excludeFilter = internalFeatureConfig.getAffectedGroupsExcludeFilter();
+        logger.debug( "Filter affected groups, exclude: {}", excludeFilter );
+        if ( isBlank( excludeFilter ) )
+        {
+            return affectedGroups;
+        }
+        return affectedGroups.stream()
+                             .filter( s -> !s.getName().matches( excludeFilter ) )
+                             .collect( Collectors.toSet() );
     }
 
 }
